+++ previous-versions/sync-for-payables-version-1/src/sdk/connections.ts
@@ -36,7 +36,7 @@ export class Connections extends ClientSDK {
    * @remarks
    * Creates a connection for the company by providing a valid `platformKey`.
    *
-   * Use the [List Integrations](https://docs.codat.io/sync-for-payables-api#/operations/list-integrations) endpoint to access valid platform keys.
+   * Use the [List Integrations](https://docs.codat.io/platform-api#/operations/list-integrations) endpoint to access valid platform keys.
    */
   async create(
     request: operations.CreateConnectionRequest,